+++ src/game/shared/singleplay_gamerules.cpp
@@ -35,7 +35,7 @@ bool CSingleplayRules::IsMultiplayer( void )
 //-----------------------------------------------------------------------------
 // Purpose: 
 //-----------------------------------------------------------------------------
-int CSingleplayRules::Damage_GetTimeBased( void )
+int64 CSingleplayRules::Damage_GetTimeBased( void )
 {
 	int iDamage = ( DMG_PARALYZE | DMG_NERVEGAS | DMG_POISON | DMG_RADIATION | DMG_DROWNRECOVER | DMG_ACID | DMG_SLOWBURN );
 	return iDamage;
@@ -44,7 +44,7 @@ int CSingleplayRules::Damage_GetTimeBased( void )
 //-----------------------------------------------------------------------------
 // Purpose: 
 //-----------------------------------------------------------------------------
-int	CSingleplayRules::Damage_GetShouldGibCorpse( void )
+int64	CSingleplayRules::Damage_GetShouldGibCorpse( void )
 {
 	int iDamage = ( DMG_CRUSH | DMG_FALL | DMG_BLAST | DMG_SONIC | DMG_CLUB );
 	return iDamage;
@@ -53,7 +53,7 @@ int	CSingleplayRules::Damage_GetShouldGibCorpse( void )
 //-----------------------------------------------------------------------------
 // Purpose: 
 //-----------------------------------------------------------------------------
-int CSingleplayRules::Damage_GetShowOnHud( void )
+int64 CSingleplayRules::Damage_GetShowOnHud( void )
 {
 	int iDamage = ( DMG_POISON | DMG_ACID | DMG_DROWN | DMG_BURN | DMG_SLOWBURN | DMG_NERVEGAS | DMG_RADIATION | DMG_SHOCK );
 	return iDamage;
@@ -62,7 +62,7 @@ int CSingleplayRules::Damage_GetShowOnHud( void )
 //-----------------------------------------------------------------------------
 // Purpose: 
 //-----------------------------------------------------------------------------
-int	CSingleplayRules::Damage_GetNoPhysicsForce( void )
+int64	CSingleplayRules::Damage_GetNoPhysicsForce( void )
 {
 	int iTimeBasedDamage = Damage_GetTimeBased();
 	int iDamage = ( DMG_FALL | DMG_BURN | DMG_PLASMA | DMG_DROWN | iTimeBasedDamage | DMG_CRUSH | DMG_PHYSGUN | DMG_PREVENT_PHYSICS_FORCE );
@@ -72,7 +72,7 @@ int	CSingleplayRules::Damage_GetNoPhysicsForce( void )
 //-----------------------------------------------------------------------------
 // Purpose: 
 //-----------------------------------------------------------------------------
-int	CSingleplayRules::Damage_GetShouldNotBleed( void )
+int64	CSingleplayRules::Damage_GetShouldNotBleed( void )
 {
 	int iDamage = ( DMG_POISON | DMG_ACID );
 	return iDamage;
@@ -84,7 +84,7 @@ int	CSingleplayRules::Damage_GetShouldNotBleed( void )
 // Input  : iDmgType - 
 // Output : Returns true on success, false on failure.
 //-----------------------------------------------------------------------------
-bool CSingleplayRules::Damage_IsTimeBased( int iDmgType )
+bool CSingleplayRules::Damage_IsTimeBased(int64 iDmgType )
 {
 	// Damage types that are time-based.
 	return ( ( iDmgType & ( DMG_PARALYZE | DMG_NERVEGAS | DMG_POISON | DMG_RADIATION | DMG_DROWNRECOVER | DMG_ACID | DMG_SLOWBURN ) ) != 0 );
@@ -95,7 +95,7 @@ bool CSingleplayRules::Damage_IsTimeBased( int iDmgType )
 // Input  : iDmgType - 
 // Output : Returns true on success, false on failure.
 //-----------------------------------------------------------------------------
-bool CSingleplayRules::Damage_ShouldGibCorpse( int iDmgType )
+bool CSingleplayRules::Damage_ShouldGibCorpse(int64 iDmgType )
 {
 	// Damage types that gib the corpse.
 	return ( ( iDmgType & ( DMG_CRUSH | DMG_FALL | DMG_BLAST | DMG_SONIC | DMG_CLUB ) ) != 0 );
@@ -106,7 +106,7 @@ bool CSingleplayRules::Damage_ShouldGibCorpse( int iDmgType )
 // Input  : iDmgType - 
 // Output : Returns true on success, false on failure.
 //-----------------------------------------------------------------------------
-bool CSingleplayRules::Damage_ShowOnHUD( int iDmgType )
+bool CSingleplayRules::Damage_ShowOnHUD(int64 iDmgType )
 {
 	// Damage types that have client HUD art.
 	return ( ( iDmgType & ( DMG_POISON | DMG_ACID | DMG_DROWN | DMG_BURN | DMG_SLOWBURN | DMG_NERVEGAS | DMG_RADIATION | DMG_SHOCK ) ) != 0 );
@@ -117,7 +117,7 @@ bool CSingleplayRules::Damage_ShowOnHUD( int iDmgType )
 // Input  : iDmgType - 
 // Output : Returns true on success, false on failure.
 //-----------------------------------------------------------------------------
-bool CSingleplayRules::Damage_NoPhysicsForce( int iDmgType )
+bool CSingleplayRules::Damage_NoPhysicsForce(int64 iDmgType )
 {
 	// Damage types that don't have to supply a physics force & position.
 	int iTimeBasedDamage = Damage_GetTimeBased();
@@ -129,7 +129,7 @@ bool CSingleplayRules::Damage_NoPhysicsForce( int iDmgType )
 // Input  : iDmgType - 
 // Output : Returns true on success, false on failure.
 //-----------------------------------------------------------------------------
-bool CSingleplayRules::Damage_ShouldNotBleed( int iDmgType )
+bool CSingleplayRules::Damage_ShouldNotBleed(int64 iDmgType )
 {
 	// Damage types that don't make the player bleed.
 	return ( ( iDmgType & ( DMG_POISON | DMG_ACID ) ) != 0 );
